first refactor using template method
