+++ config/constants.py
@@ -5,7 +5,7 @@ class Constants(object):
     normalMonths = ['11','04','06', '09']
     shortMonths = ['02']
     startYear = 2020
-    endYear = 2020
+    endYear = 2100
     numMonths = 12
     numDaysShortMonth = 28
     numDaysNormalMonth = 30
