+++ service/date/i18nDate.py
@@ -1,0 +1,7 @@
+#!/usr/bin/python3
+
+from service.date.date import AbstractDate
+
+class I18NDate(AbstractDate):
+    def compileDate(self, year, month, day):
+        return year + '-' + month + '-' + day
