+++ service/date/itDate.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python3
+
+from service.date.date import AbstractDate
+
+class ITDate(AbstractDate):
+    def compileDate(self, year, month, day):
+        return day + '-' + month + '-' + year
+
